Test-set pipeline, drift analysis, and float32 memory fixes
Test set is now fully through the pipeline: 308.7M raw rows uploaded and staged
with exact row counts (market 118,359,166 / order 119,465,287 / transaction
70,909,408), then the SAME feature SQL as train produced dataset_test with
647,896 rows x 295 columns. submission.csv validates against the Kaggle template:
identical sample_id set, no NaN, prediction std 0.000298 (0.114x the target std,
the expected shrinkage of a regression estimator - and irrelevant to a
scale-invariant metric).

src/evaluation/drift.py - train vs test distribution drift by effect size rather
than p-value, because at 1.26M vs 648k samples every distributional test returns
a vanishing p-value for a difference too small to matter.

Findings (294 features): 201 negligible, 85 small, 8 moderate, 0 large.
Median shift 0.028, max 0.348. The split between feature types is sharp:
  scale-free (imbalance, OFI, returns)  median shift 0.002 - 0.007
  rate / count                          median shift 0.176, max 0.348
This validates the design choice to build imbalances rather than raw counts, and
it converges with SHAP: the most predictive features are also the most transferable.
The honest caveat is that rate features are still used and they do shift.

The test period is a measurably more liquid regime: 1.36x order events, 1.35x
trade count, 1.24x depth, but 0.80x spread and 0.91x mid volatility.

Memory fixes (both were real failures on this 16 GB machine):
- load_dataset() now loads float32, halving peak memory from 2.96 GB to 1.48 GB.
  The float64 path raised MemoryError while another job held 2 GB.
- Predictor._frame and the submission step no longer force float64. Generating the
  test submission needed 1.4 GB as float64 and died; float32 needs 0.7 GB.
Nothing is lost: source columns are float32, tree models bin to uint8 internally,
and metrics still promote to float64.

95 tests pass.

+++ src/inference/predictor.py
@@ -118,7 +118,11 @@ class Predictor:
         extra = [c for c in df.columns if c not in expected]
         if extra:
             log.warning("ignored %d unexpected field(s) in request: %s", len(extra), extra[:5])
-        return df[expected].astype("float64")
+        # float32, not float64: LightGBM and XGBoost accept it natively and bin to
+        # uint8 internally, so nothing is lost - but a 648k x 289 batch needs 1.4 GB
+        # as float64 versus 0.7 GB as float32. The float64 version ran out of memory
+        # while generating the test submission on a 16 GB machine.
+        return df[expected].astype("float32")
 
     def predict(self, rows: list[dict]) -> np.ndarray:
         if not rows:
